Increase the minimum difference between the upper and lower thresholds to 10.
Having this at 5 is causing confusion, since the hardware isn't actually designed
to have the thresholds so close together.  This was set to 5 for internal testing,
but it's not actually useful to have them so close together.  The thresholds exist
to prevent ghosting and strobing on the panels, and putting the thresholds very
close together prevents them from doing that.