Adjust the high FSR preset.
The 100-112 thresholds are extremely low, and don't work for all play styles, since
with harder presses, the 100 low threshold may not be reached quickly enough, which
causes missed inputs.  Raise the high preset thresholds to something more reasonable.
Users can still set the lower values in the advanced tab if it works for them.
This might be tuned further.

This won't affect users who are already set to the higher thresholds, since it'll
just act like custom thresholds.  To update thresholds to the new values, click the
"High" preset button again.